[change] move crud operations to their own module
