+++ ata_api/main.py
@@ -1,16 +1,16 @@
 import random
-from typing import Annotated, Generator, Optional
+from typing import Annotated, Generator
 from uuid import UUID
 
 from ata_db_models.helpers import get_conn_string
-from ata_db_models.models import Group, UserGroup
-from fastapi import Depends, FastAPI, HTTPException, Path, Query, status
+from ata_db_models.models import Group
+from fastapi import Depends, FastAPI, Path, Query
 from mangum import Mangum
 from pydantic import BaseModel
 from sqlalchemy.orm import Session, sessionmaker
-from sqlmodel import create_engine, select
+from sqlmodel import create_engine
 
-from ata_api.helpers.decorators import raise_exception
+from ata_api.crud import create_prescription, get_prescription
 from ata_api.helpers.enums import SiteName
 from ata_api.helpers.logging import logging
 
@@ -42,36 +42,6 @@ class PrescriptionResponse(BaseModel):
 @app.get("/")
 def get_root() -> object:
     return {"message": "This is the root endpoint for the AtA API."}
-
-
-@raise_exception(
-    HTTPException(
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        detail="An exception occurred while fetching the prescription.",
-    )
-)
-def get_prescription(session: Session, site_name: SiteName, user_id: UUID) -> Optional[UserGroup]:
-    result = session.execute(
-        select(UserGroup).where(UserGroup.user_id == user_id, UserGroup.site_name == site_name)
-    ).one_or_none()
-    return result[0] if result is not None else None
-
-
-@raise_exception(
-    HTTPException(
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        detail="An exception occurred while creating the prescription.",
-    )
-)
-def create_prescription(session: Session, site_name: SiteName, user_id: UUID, group: Group) -> UserGroup:
-    usergroup = UserGroup(
-        site_name=site_name,
-        user_id=user_id,
-        group=group,
-    )
-    session.add(usergroup)
-    session.commit()
-    return usergroup
 
 
 @app.get("/prescription/{site_name}/{user_id}", response_model=PrescriptionResponse)
